reset deadline before socket read, refactor, fix linter warning
This commit does the following:

* Fixes a linter warning by directly returning an `error` result
instead of checking it for nilness.
* Refactors all of the functions except one helper to make them
methods of the `Conn` object, thus giving them access to the
timeout and the underlying `net.Conn` object to be able to use
`SetDeadline`.
* Sets a new deadline before each socket read in `readFrame()`,
preventing timeouts between and during frame reads.
* Makes the maximum frame body size an exported, package-level
variable as the previously set body size was only 1 MiB,
Bitcoin blocks can be 4MB, and max body size is 2^63-1 as
defined by ZMTP RFC. The default is set to 32MiB, the maximum
Bitcoin message size.